back_end/src/gateway/requirements.txt
flask
flask_cors
flask-socketio
requests
pika
flask_jwt_extended